refactor(extensions): move extension functions to new folder
